pyproject.toml
[project]
name = "documentation-mcp-server"
version = "0.1.0"
description = "MCP server that indexes documentation from git repos and serves it to AI agents"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.22.0,<2",
    "chromadb>=1.5.0,<2",
    "sentence-transformers>=5.0.0,<6",
    "gitpython>=3.1.41,<4",
    "pyyaml>=6.0",
    "apscheduler>=3.10.0,<4",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/docserver"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=docserver --cov-report=term-missing"

[tool.coverage.run]
source = ["src/docserver"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if __name__",
    "if TYPE_CHECKING",
]
